requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic[email]
python-dotenv==1.0.0
python-decouple==3.8
PyJWT==2.8.0
bcrypt==4.0.1
cryptography==41.0.5
email-validator==2.1.1